+++ formula/src/main/java/com/instacart/formula/StatelessFormula.kt
@@ -23,6 +23,8 @@ abstract class StatelessFormula<Input, Output> : IFormula<Input, Output> {
         ): Evaluation<Output> {
             return evaluate(input, context)
         }
+
+        override fun key(input: Input): Any? = this@StatelessFormula.key(input)
     }
 
     /**
@@ -41,6 +43,15 @@ abstract class StatelessFormula<Input, Output> : IFormula<Input, Output> {
         input: Input,
         context: FormulaContext<Unit>
     ): Evaluation<Output>
+
+    /**
+     * A unique identifier used to distinguish formulas of the same type.
+     *
+     * ```
+     * override fun key(input: ItemInput) = input.itemId
+     * ```
+     */
+    open fun key(input: Input): Any? = null
 
     override fun implementation(): Formula<Input, *, Output> = implementation
 }