+++ formula/src/test/java/com/instacart/formula/DuplicateChildrenTest.kt
@@ -11,10 +11,9 @@ class DuplicateChildrenTest {
         }
     }
 
-    class ParentFormula : Formula<Unit, Unit, List<Unit>> {
-        override fun initialState(input: Unit) = Unit
+    class ParentFormula : StatelessFormula<Unit, List<Unit>>() {
 
-        override fun evaluate(input: Unit, state: Unit, context: FormulaContext<Unit>): Evaluation<List<Unit>> {
+        override fun evaluate(input: Unit, context: FormulaContext<Unit>): Evaluation<List<Unit>> {
             return Evaluation(
                 output = listOf(1, 2, 3).map {
                     context.child(ChildFormula())
@@ -23,10 +22,8 @@ class DuplicateChildrenTest {
         }
     }
 
-    class ChildFormula: Formula<Unit, Unit, Unit> {
-        override fun initialState(input: Unit) = Unit
-
-        override fun evaluate(input: Unit, state: Unit, context: FormulaContext<Unit>): Evaluation<Unit> {
+    class ChildFormula: StatelessFormula<Unit, Unit>() {
+        override fun evaluate(input: Unit, context: FormulaContext<Unit>): Evaluation<Unit> {
             return Evaluation(
                 output = Unit
             )